ci: integrate GitHub Action for DB dumps and Laravel app bundles (#70)
Integrate a GitHub Action to facilitate the installation on shared
hosting solutions. This action automates the upload of MariaDB and
PostgreSQL database dumps, along with zip archives containing the
application bundled with Composer packages. The action is triggered on
the push of a git tag, enhancing the deployment workflow for shared
hosting environments.

Signed-off-by: Valentin Sickert <17144397+Lapotor@users.noreply.github.com>